+++ packages/stylelint-config-wordpress/scss.js
@@ -12,9 +12,7 @@ module.exports = {
 	rules: {
 
 		// stylelint-config-wordpress css overrides
-		'at-rule-no-unknown': [ true, {
-			ignoreAtRules: [ 'extend', 'at-root', 'warn', 'error', 'if', 'else', 'for', 'each', 'while', 'mixin', 'include', 'content', 'return', 'function' ],
-		} ],
+		'at-rule-no-unknown': null,
 		'at-rule-empty-line-before': [ 'always', {
 			except: ['blockless-after-blockless'],
 			ignore: ['after-comment'],
@@ -31,6 +29,7 @@ module.exports = {
 		'scss/at-else-closing-brace-newline-after': 'always-last-in-chain',
 		'scss/at-else-closing-brace-space-after': 'always-intermediate',
 		'scss/at-else-empty-line-before': 'never',
+		'scss/at-rule-no-unknown': true,
 		'scss/at-if-closing-brace-newline-after': 'always-last-in-chain',
 		'scss/at-if-closing-brace-space-after': 'always-intermediate',
 		'scss/selector-no-redundant-nesting-selector': true,