+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -1,6 +1,9 @@
 # HEAD
 
 -   Changed: `stylelint-config-wordpress` now extends [`stylelint-config-recommended`](https://github.com/stylelint/stylelint-config-recommended), which turns on the `at-rule-no-unknown`, `block-no-empty`, `comment-no-empty`, `declaration-block-no-ignored-properties`, `declaration-block-no-redundant-longhand-properties`, `font-family-no-duplicate-names`, `media-feature-name-no-unknown`, `no-empty-source` rule. These rules are part of stylelint's [possible errors](https://github.com/stylelint/stylelint/blob/master/docs/user-guide/rules.md#possible-errors) rules.
+-   Removed: `stylelint-scss < 1.5.1` compatibility.
+-   Removed: `at-rule-no-unknown` custom `ignoreAtRules` options in `stylelint-config-wordpress/scss` shared config.
+-   Added: `scss/at-rule-no-unknown` rule in `stylelint-config-wordpress/scss` shared config.
 
 # 11.0.0
 
@@ -18,7 +21,7 @@
 
 # 10.0.0
 
--   Added: `scss/selector-no-redundant-nesting-selector` rule in `stylelint-config-wordpress/scss` chared config.
+-   Added: `scss/selector-no-redundant-nesting-selector` rule in `stylelint-config-wordpress/scss` shared config.
 -   Added: `selector-no-empty` rule.
 -   Added: NodeJS 7.x.x support
 -   Fixed: Added `stylelint-scss` plugin @if/@else placement rules.